refactor(apps/bridge): crossChainLiquidityPositions
