+++ packages/graph-client/resolvers/sushiswap/index.ts
@@ -7,4 +7,4 @@ export * from './crossChainPair'
 export * from './crossChainPairs'
 export * from './crossChainToken'
 export * from './crossChainTokens'
-export * from './crossChainUsers'
+export * from './crossChainLiquidityPositions'
